% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdth_catches_sp_fishingmode_year_fleet_ocean.R
\name{avdth_catches_sp_fishingmode_year_fleet_ocean}
\alias{avdth_catches_sp_fishingmode_year_fleet_ocean}
\title{Catches by species, fishing mode, year, fleet and ocean}
\usage{
avdth_catches_sp_fishingmode_year_fleet_ocean(avdth_con, year, fleet,
  ocean, fishing_mode, fleet_name, specie = c("BET", "YFT", "SKJ"))
}
\arguments{
\item{avdth_con}{AVDTH database connection object.}

\item{year}{Year selected (numeric value). You can select only one year (related to output design).}

\item{fleet}{Fleet(s) selected (numeric value). You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{Ocean selected (numeric value). You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fishing_mode}{Type of fishing mode. Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{Fleet(s) name(s) (character value).}

\item{specie}{Specie(s) name(s) selected. Specify specie code (on 3 letters) and add several species with the function c(). If you want to display all the species available, enter "all" in the argument. By default the function shows the 3 major tropical tunas (YFT, BET and SKJ).}
}
\value{
A R list with data/informations for produce a graphic (stacked area) associated to query data specifications.
}
\description{
Catches by species, fishing mode, year, fleet and ocean.
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean
# For the argument fishing_mode, 1 = Floating object
# By default, for 3 major tropical tunas
tmp1 <- avdth_catches_sp_fishingmode_year_fleet_ocean(avdth_con = avdth_connection,
                                                      year = 2017,
                                                      fleet = c(1, 41),
                                                      ocean = 1,
                                                      fishing_mode = 1
                                                      fleet_name = "french fleet")
# For display all the species available
tmp2 <- avdth_catches_sp_fishingmode_year_fleet_ocean(avdth_con = avdth_connection,
                                                      year = 2017,
                                                      fleet = c(1, 41),
                                                      ocean = 1,
                                                      fishing_mode = 1
                                                      fleet_name = "french fleet",
                                                      specie = "all")
# For specify one (or more species), here with the example of Thunnus alalunga (ALB) and Auxis rochei (BLT)
tmp3 <- avdth_catches_sp_fishingmode_year_fleet_ocean(avdth_con = avdth_connection,
                                                      year = 2017,
                                                      fleet = c(1, 41),
                                                      ocean = 1,
                                                      fishing_mode = 1
                                                      fleet_name = "french fleet",
                                                      specie = c("ALB", "BLT"))
}
\references{
\url{https://github.com/OB7-IRD/fishi}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
