% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdth_nb_landing_fleet_ocean.R
\name{avdth_nb_landing_year_fleet_ocean}
\alias{avdth_nb_landing_year_fleet_ocean}
\title{Number of landings by year, fleet and ocean}
\usage{
avdth_nb_landing_year_fleet_ocean(avdth_con, year, fleet, ocean,
  fleet_name)
}
\arguments{
\item{avdth_con}{A AVDTH database connection object.}

\item{year}{Year selected (numeric value). You can select only one year (related to output design).}

\item{fleet}{Fleet(s) selected (numeric value). You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{Ocean selected (numeric value). You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{Fleet(s) name(s) (character value).}
}
\value{
One graphic associated to query data specifications.
}
\description{
Number of landings by year, fleet and ocean.
}
\references{
\url{https://github.com/OB7-IRD/fishi}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
