% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdth_nblanding_year_month_fleet_ocean.R
\name{avdth_nblanding_year_month_fleet_ocean}
\alias{avdth_nblanding_year_month_fleet_ocean}
\title{Number of landings by year, month, fleet and ocean (associated to an AVDTH database)}
\usage{
avdth_nblanding_year_month_fleet_ocean(
  avdth_con,
  year,
  fleet,
  ocean,
  fleet_name
)
}
\arguments{
\item{avdth_con}{(JDBCConnection object) AVDTH database connection object.}

\item{year}{(integer) Year selected. You can select only one year (related to output design).}

\item{fleet}{(integer) Fleet(s) selected. You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{(integer) Ocean selected. You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{(character) Fleet(s) name(s).}
}
\value{
A ggplot object.
}
\description{
Number of landings by year, month, fleet and ocean (associated to an AVDTH database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean
\dontrun{
tmp <- avdth_nblanding_year_fleet_ocean(avdth_con = avdth_connection,
                                        year = as.integer(2017),
                                        fleet = as.integer(c(1 , 41)),
                                        ocean = as.integer(1),
                                        fleet_name = "french fleet")}
}
