% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/avdth_nbset_fishingmode_year_month_fleet_ocean.R
\name{avdth_nbset_fishingmode_year_month_fleet_ocean}
\alias{avdth_nbset_fishingmode_year_month_fleet_ocean}
\title{Number of set by fishing mode, year, month, fleet and ocean (associated to an AVDTH database)}
\usage{
avdth_nbset_fishingmode_year_month_fleet_ocean(avdth_con, year, fleet,
  ocean, fleet_name)
}
\arguments{
\item{avdth_con}{AVDTH database connection object.}

\item{year}{Year selected (numeric value). You can select only one year (related to output design).}

\item{fleet}{Fleet(s) selected (numeric value). You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{Ocean selected (numeric value). You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{Fleet(s) name(s) (character value).}
}
\value{
A R list with data/informations for produce a graphic (grouped barplot) associated to query data specifications.
}
\description{
Number of set by fishing mode, year, month, fleet and ocean (associated to an AVDTH database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean
tmp <- avdth_nbset_fishingmode_year_month_fleet_ocean(avdth_con = avdth_connection,
                                                      year = 2017,
                                                      fleet = c(1 , 41),
                                                      ocean = 1,
                                                      fleet_name = "french fleet")
}
\references{
\url{https://github.com/OB7-IRD/fishi}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
