% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/avdth_setloc_fishingmode_year_month_fleet_ocean.R
\name{avdth_setloc_fishingmode_year_month_fleet_ocean}
\alias{avdth_setloc_fishingmode_year_month_fleet_ocean}
\title{Set location by year, month, fleet, fishing mode and ocean (associated to an AVDTH database)}
\usage{
avdth_setloc_fishingmode_year_month_fleet_ocean(
  avdth_con,
  year,
  fleet,
  ocean,
  fishing_mode,
  fleet_name
)
}
\arguments{
\item{avdth_con}{(JDBCConnection object) AVDTH database connection object.}

\item{year}{(integer) Year selected. You can select only one year (related to output design).}

\item{fleet}{(integer) Fleet(s) selected. You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{(integer) Ocean selected. You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fishing_mode}{(integer) Type of fishing mode. Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{(character) Fleet(s) name(s).}
}
\value{
A ggplot object.
}
\description{
Set location by year, month, fleet, fishing mode and ocean (associated to an AVDTH database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean
# For the argument fishing_mode, 1 = Floating object
\dontrun{
tmp <- avdth_setloc_fishingmode_year_month_fleet_ocean(avdth_con = avdth_connection,
                                                       year = 2017,
                                                       fleet = c(1 , 41),
                                                       ocean = 1,
                                                       fishing_mode = 1,
                                                       fleet_name = "french fleet")}
}
