% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/balbaya_nbset_fishingmode_gear_year_month_fleet_ocean.R
\name{balbaya_nbset_fishingmode_gear_year_month_fleet_ocean}
\alias{balbaya_nbset_fishingmode_gear_year_month_fleet_ocean}
\title{Number of sets by fishing mode, gear, year, month, fleet and ocean (associated to a balbaya database)}
\usage{
balbaya_nbset_fishingmode_gear_year_month_fleet_ocean(
  balbaya_con,
  year,
  fleet,
  ocean,
  fishing_mode,
  gear,
  fleet_name,
  monthly = TRUE,
  neg_set = TRUE,
  pos_set = TRUE,
  acronym = TRUE
)
}
\arguments{
\item{balbaya_con}{(JDBCConnection object) Balbaya database connection object.}

\item{year}{(integer) Year(s) selected}

\item{fleet}{(integer) Fleet(s) selected.}

\item{ocean}{(integer) Ocean(s) selected.}

\item{fishing_mode}{(integer) Type(s) of fishing mode.}

\item{gear}{Gear name(s).}

\item{fleet_name}{(character) Fleet name(s).}

\item{monthly}{(logical) Display information monthly. By default TRUE}

\item{neg_set}{(logical) Display information about number of negative set. By defaut TRUE.}

\item{pos_set}{(logical) Display information about number of postive set. By defaut TRUE.}

\item{acronyme}{(logical) Show acronym in the legend or full term. Be default TRUE.}
}
\value{
A ggplot object.
}
\description{
Number of sets by fishing mode, year, month, fleet and ocean (associated to a balbaya database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean and 2 = Indian Ocean
# For the argument fishing_mode, 1 = Floating object and 2 = Free school
# For the argument gear, 1 = Purse seiner
\dontrun{
tmp1 <- balbaya_nbset_fishingmode_gear_year_month_fleet_ocean(balbaya_con = balabay_connection,
                                                              year = c(2017, 2018),
                                                              fleet = c(1, 41),
                                                              ocean = c(1, 2),
                                                              fishing_mode = c(1, 2),
                                                              gear = 1,
                                                              fleet_name = "French fleet",
                                                              monthly = TRUE,
                                                              neg_set = TRUE,
                                                              pos_set = TRUE,
                                                              acronym = TRUE)}
}
