% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean.R
\name{balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean}
\alias{balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean}
\title{Temporal variables of effort by fishing mode, gear, year, month, fleet and ocean (associated to a balbaya database)}
\usage{
balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean(
  balbaya_con,
  year,
  fleet,
  ocean,
  fishing_mode,
  gear,
  fleet_name,
  monthly = TRUE,
  effort_variable = c("time_at_sea", "days_at_sea", "fishing_time", "fishing_days",
    "searching_days"),
  acronym = TRUE
)
}
\arguments{
\item{balbaya_con}{(JDBCConnection object) Balbaya database connection object.}

\item{year}{(interger) Year(s) selected.}

\item{fleet}{(integer) Fleet(s) selected.}

\item{ocean}{(integer) Ocean(s) selected.}

\item{fishing_mode}{(integer) Type(s) of fishing mode.}

\item{gear}{(integer) Gear name(s).}

\item{fleet_name}{(character) Fleet name(s).}

\item{monthly}{(logical) Display information monthly. By default TRUE}

\item{variables}{(character) List of variables display in the graphic. By default all variables availabe will display.}

\item{acronyme}{Show acronym in the legend or full term. By default TRUE.}
}
\value{
A ggplot object.
}
\description{
Temporal variables of effort by fishing mode, year, month, fleet and ocean (associated to a balbaya database).
}
\details{
For now, you can display 5 temporal variables:
\itemize{
 \item{time_at_sea: }{time (in hours) of the vessel at sea.}
 \item{days_at_sea: }{time (in days) of the vessel at sea.}
 \item{fishing_time: }{time (in hours) of vessel fishing activity.}
 \item{fishing_days: }{time (in days) of vessel fishing activity. In the Atlantic Ocean a full day of activity is 12 hours by day and 13 hours for the Indian Ocean.}
 \item{searching_days: }{time (in days) of vessel fishing activity (variable fishing_time) less the set duration (time required to make a catch, start when the fishes is incircled).}
}
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean and 2 = Indian Ocean
# For the argument fishing_mode, 1 = Floating object and 2 = Free school
# For the argument gear, 1 = Purse seiner
\dontrun{
tmp1 <- balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean(balbaya_con = balabay_connection,
                                                                        year = c(2017, 2018),
                                                                        fleet = c(1, 41),
                                                                        ocean = c(1, 2),
                                                                        fishing_mode = c(1, 2),
                                                                        gear = 1,
                                                                        fleet_name = "French fleet",
                                                                        monthly = TRUE,
                                                                        variables = c("time_at_sea", "days_at_sea", "fishing_time", "fishing_days", "searching_days"),
                                                                        acronym = TRUE)}
}
