% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_size_tuna.R
\name{bio_size_tuna}
\alias{bio_size_tuna}
\title{Bio size tuna}
\usage{
bio_size_tuna(
  data_connection,
  report_year,
  country = as.integer(x = 1),
  ocean = as.integer(x = 1),
  graph_type = "plot"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the bio_size_tuna() function.}

\item{report_year}{{\link[base]{integer}} expected. Year of the statistical report.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{graph_type}{{\link[base]{character}} expected. plot or plotly. Plot by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Size distribution of major tuna catches (in percentage of the total number of fishes) for the French purse seine fleet.
}
