% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_weight_tuna.R
\name{bio_weight_tuna}
\alias{bio_weight_tuna}
\title{Bio weight tuna}
\usage{
bio_weight_tuna(
  data_connection,
  report_year,
  ocean,
  country = as.integer(x = 1),
  graph_type = "plot"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the bio_weight_tuna() function.}

\item{report_year}{{\link[base]{integer}} expected. Year of the statistical report.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification. 1 by default.}

\item{graph_type}{{\link[base]{character}} expected. plot or ggplot. Plot by default.}
}
\value{
The function return ggplot R plot.
}
\description{
weight distribution of major tuna catches (in percentage of the total number of fishes) for the French purse seine fleet.
}
