% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_UO_count.R
\name{biological_UO_count}
\alias{biological_UO_count}
\title{Number of biological UO}
\usage{
biological_UO_count(
  path_file,
  graph_type = "number",
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{path_file}{{\link[base]{character}} expected. File path to the Tunabio file.}

\item{graph_type}{{\link[base]{character}} expected. "number" or "table." Number give the total number of each biological UO and table give the detail per month. Number by default.}

\item{start_date}{{\link[base]{date}} expected. Write the start date of the contract}

\item{end_date}{{\link[base]{date}} expected. Write the end date of the contract}
}
\value{
The function return table.
}
\description{
Count the UO for biological sampling.
}
