% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_per_positive_set.R
\name{catch_per_positive_set}
\alias{catch_per_positive_set}
\title{Annual number of catch per positive set}
\usage{
catch_per_positive_set(
  dataframe,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the query named "balbaya_catch_effort.sql".}

\item{fishing_type}{{\link[base]{character}} expected. 'FOB' and 'FSC'.}

\item{graph_type}{{\link[base]{character}} expected. 'plot', 'plotly' or 'table'. Plot by default.}

\item{title}{TRUE or FALSE expected. Title for plotly graph_type. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of catch per positive set on FOB-associated and free-swimming schools.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:

Dataframe:
\preformatted{
   activity_date | species_code | school_type | set_duration | positive_set | total_set | total_catch_weight | total_hour_fished
   -------------------------------------------------------------------------------------------------------------------------------
   1999-07-09    | 2            | FOB         | 3.54         | 1            | 1         | 119.0              | 12.1
   1999-07-09    | 1            | FOB         | 3.54         | 1            | 1         | 20.6               | 12.1
   1999-07-09    | 1            | FOB         | 3.54         | 1            | 1         | 24.4               | 12.1
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
 \item{\code{  vessel_type_code}}
}
}
