% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_per_unit_effort.R
\name{catch_per_unit_effort}
\alias{catch_per_unit_effort}
\title{Annual catch rates (in t per searching day)}
\usage{
catch_per_unit_effort(
  dataframe1,
  dataframe2,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{fishing_type}{{\link[base]{character}} expected. FOB or FSC.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual catch rates (in t per searching day) on FOB- associated and free-swimming tuna schools (FSC).
}
