% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_serie.R
\name{catch_serie}
\alias{catch_serie}
\title{Catch serie}
\usage{
catch_serie(
  data_connection,
  time_period,
  specie,
  ocean,
  country,
  vessel_type,
  time_step = "year",
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the catch_serie function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year, between 1981 and 2021.}

\item{specie}{{\link[base]{integer}} expected. Select the specie : 1 YFT, 2 SKJ, 3 BET, 4 ALB (cf the vignette "referentials" for more).}

\item{ocean}{{\link[base]{integer}} expected. Select the ocean : 1 Atlantic, 2 Indian, 3 West Pacific, 4 East Pacific, 5 Pacific.}

\item{country}{{\link[base]{integer}} expected. Select the country : 1 France, 41 Mayotte}

\item{vessel_type}{{\link[base]{integer}} expected. Select the vessel type : 1 purse seiner, 2 bait boat, 3 longliner, 4 support, 5 mix.}

\item{time_step}{{\link[base]{character}} expected. Parameter used in the graphical representation. By year by default, to represent by month specify "month".}

\item{path_file}{path to save the final graphic as a png. NULL by default.}
}
\value{
The function return  ggplot R object.
}
\description{
The purpose of the catch_serie function is to graphically represent the evolution of catches over time by a barplot. It depends on several arguments: the period of time, the species, the oceans, the countries and the types of vessels previously selected by the user. The representation can be done by year or by month depending on the wished result.
}
