% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_acronym}
\alias{check_acronym}
\title{Arugment "acronym" verification}
\usage{
check_acronym(acronym)
}
\arguments{
\item{acronym}{(logical) Information display on the graph with acronym(s)}
}
\description{
Check if the item "acronym" is provided and is logical value
}
