% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_avdth_con}
\alias{check_avdth_con}
\title{Arugment "avdth_con" verification}
\usage{
check_avdth_con(avdth_con)
}
\arguments{
\item{avdth_con}{(JDBCConnection object) Connection object for a AVDTH database}
}
\description{
Check if the item "avdth_con" is provided and have one unique class identified as JDBCConnection
}
