% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_balbaya_con}
\alias{check_balbaya_con}
\title{Arugment "balbaya_con" verification}
\usage{
check_balbaya_con(balbaya_con)
}
\arguments{
\item{balbaya_con}{(PostgreSQLConnection object) Connection object for a balbaya database}
}
\description{
Check if the item "balbaya_con" is provided and have one unique class identified as PostgreSQLConnection
}
