% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_effort_variable}
\alias{check_effort_variable}
\title{Arugment "effort_variable" verification}
\usage{
check_effort_variable(effort_variable)
}
\arguments{
\item{effort_variable}{(character) Effort's variable(s) display}
}
\description{
Check if the item "effort_variable" is provided, have a class identified as character
}
