% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fishing_mode}
\alias{check_fishing_mode}
\title{Arugment "fishing_mode" verification}
\usage{
check_fishing_mode(fishing_mode, several_values)
}
\arguments{
\item{fishing_mode}{(integer) Fishing mode identification(s)}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "fishing_mode" is provided, is integer class and have one or several values.
}
