% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fleet}
\alias{check_fleet}
\title{Arugment "fleet" verification}
\usage{
check_fleet(fleet, several_values)
}
\arguments{
\item{fleet}{(integer) Fleet identification(s)}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "fleet" is provided and interger value(s)
}
