% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fleet_name}
\alias{check_fleet_name}
\title{Arugment "fleet_name" verification}
\usage{
check_fleet_name(fleet_name)
}
\arguments{
\item{fleet_name}{(character) Fleet name identification}
}
\description{
Check if the item "fleet_name" is provided and have a class identified as character
}
