% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gear}
\alias{check_gear}
\title{Arugment "gear" verification}
\usage{
check_gear(gear, several_values)
}
\arguments{
\item{gear}{(integer) Gear(s) identification}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "gear" is provided and have integer value(s)
}
