% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_monthly}
\alias{check_monthly}
\title{Arugment "monthly" verification}
\usage{
check_monthly(monthly)
}
\arguments{
\item{monthly}{(logical) Display information monthly.}
}
\description{
Check if the item "monthly" is provided and is logical value
}
