% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_neg_set}
\alias{check_neg_set}
\title{Arugment "neg_set" verification}
\usage{
check_neg_set(neg_set)
}
\arguments{
\item{neg_set}{(logical) Display information about negative set.}
}
\description{
Check if the item "neg_set" is provided and is logical value
}
