% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ocean}
\alias{check_ocean}
\title{Arugment "ocean" verification}
\usage{
check_ocean(ocean, several_values)
}
\arguments{
\item{ocean}{(integer) Ocean identification}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "ocean" is provided and have integer value(s)
}
