% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_percentage}
\alias{check_percentage}
\title{Arugment "percentage" verification}
\usage{
check_percentage(percentage)
}
\arguments{
\item{percentage}{(logical) Values display in percentage or not}
}
\description{
Check if the item "percentage" is provided and is logical value
}
