% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_pos_set}
\alias{check_pos_set}
\title{Arugment "pos_set" verification}
\usage{
check_pos_set(pos_set)
}
\arguments{
\item{pos_set}{(logical) Display information about postive set.}
}
\description{
Check if the item "pos_set" is provided and is logical value
}
