% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_specie}
\alias{check_specie}
\title{Arugment "specie" verification}
\usage{
check_specie(specie, several_values)
}
\arguments{
\item{specie}{(character) Specie name(s) coded on 3 letters}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "specie" is provided, have a class identified as character and coded on 3 letters
}
