% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_year}
\alias{check_year}
\title{Arugment "year" verification}
\usage{
check_year(year)
}
\arguments{
\item{year}{(integer) Year selected}
}
\description{
Check if the item "year" is provided and have a unique value identified as class integer
}
