% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{data_extraction}
\alias{data_extraction}
\title{Data extraction}
\usage{
data_extraction(
  type,
  data_connection = NULL,
  database_name = NULL,
  sql_name = NULL,
  csv_path = NULL,
  time_period = NULL,
  country = NULL,
  vessel_type = NULL,
  vessel_type_select = NULL,
  ocean = NULL
)
}
\arguments{
\item{type}{{\link[base]{character}} expected. dabase or csv.}

\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishing_effort function.}

\item{database_name}{{\link[base]{character}}. Name of the database.}

\item{sql_name}{{\link[base]{character}}. Name of the sql file, in fishi.}

\item{csv_path}{{\link[base]{character}}. Path of the csv.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{vessel_type_select}{{\link[base]{character}} expected. engin or vessel_type.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}
}
\value{
The function return ggplot R plot.
}
\description{
Extracts a dataset from a database or csv
}
