% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{dbconnection_control}
\alias{dbconnection_control}
\title{Control for argument dbconnection}
\usage{
dbconnection_control(dbconnection)
}
\description{
Control for argument dbconnection
}
\keyword{internal}
