% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_map.R
\name{effort_map}
\alias{effort_map}
\title{Effort map}
\usage{
effort_map(
  data_connection,
  time_period,
  ocean,
  country,
  vessel_type,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the catch_serie function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year, between 1981 and 2021.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{path_file}{path to save the final graphic as a png. NULL by default.}
}
\value{
The function return  ggplot R map.
}
\description{
The purpose of the effort_map function is to graphically represent the effort throught the oceans. It depends on several arguments: the period of time, the oceans, the countries and the types of vessels previously selected by the user.
}
