% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_production.R
\name{fishery_production}
\alias{fishery_production}
\title{Total fishery production}
\usage{
fishery_production(
  data_connection,
  time_period,
  ocean,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  vessel_type_select = "engin",
  fishing_type = "ALL",
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishery_production() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification. 1 by default.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification. 1 by default.}

\item{vessel_type_select}{{\link[base]{character}} expected. engin or vessel_type.}

\item{fishing_type}{{\link[base]{character}} expected. FSC, FOB or ALL.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly, table or percentage. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
total fishery production (catch by species).
}
