% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_effort.R
\name{fishing_effort}
\alias{fishing_effort}
\title{Annual total number of fishing and searching days}
\usage{
fishing_effort(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_capacity function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in nominal effort over time. Annual total number of fishing and searching days.
}
