% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_time.R
\name{fishing_time}
\alias{fishing_time}
\title{Fleet Operational Hours}
\usage{
fishing_time(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the fishing_time() function.}

\item{graph_type}{{\link[base]{character}} expected. 'plot', 'plotly' or 'table'. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Represents the time spent by the fleet (in hours)
}
