% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_catch_fad_previous.R
\name{map_catch_fad_previous}
\alias{map_catch_fad_previous}
\title{Map catch fad previous year}
\usage{
map_catch_fad_previous(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = 1
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the map_catch_fad_previous() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{fishing_type}{{\link[base]{character}} expected. FOB or FSC.}
}
\value{
The function return ggplot R plot.
}
\description{
Spatial distribution of tuna catches of the French purse seine fishing fleet made on FOB-associated schools.
}
