% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_effort_distribution.R
\name{map_effort_distribution}
\alias{map_effort_distribution}
\title{Spatial distribution of tuna effort}
\usage{
map_effort_distribution(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the map_effort_distribution() function.}

\item{graph_type}{{\link[base]{character}} expected. 'plot' or 'plotly.' Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Spatial distribution of tuna effort.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   activity_date | cwp11_act | ocean_code | set_duration | total_hour_fished
   --------------------------------------------------------------------------
   2022-01-02    | 301000    | 1          | 2.61         |  6.17
   2022-01-02    | 301000    | 1          | 2.61         |  6.17
   2022-01-02    | 301000    | 1          | 2.42         |  5.98
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  vessel_type_code}}
}
}
