% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_fad.R
\name{number_fad}
\alias{number_fad}
\title{Number fads}
\usage{
number_fad(dataframe, time_period = as.integer(2013:2022), table = "by_year")
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the fishery_production() function.}

\item{time_period}{{\link[base]{integer}} expected. as.integer(2013:2022) by default.}

\item{table}{{\link[base]{character}} expected. 'by_year' or 'by_vessel'. 'by_year' by default.}
}
\value{
The function return a table.
}
\description{
Number fads by year
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   fishing_year | object_code_activity | buoy_code_activity | buoy_id | vessel_type | vessel_name
   ----------------------------------------------------------------------------------------------
   1999         | 1                    | 1                  | 99972   | PS          | AVEL VAD
   1999         | 1                    | 1                  | 99971   | PS          | AVEL VAD
   1999         | 1                    | 1                  | 99970   | PS          | AVEL VAD
}
}
