% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_vessel_sampled.R
\name{number_vessel_sampled}
\alias{number_vessel_sampled}
\title{Vessel sampled}
\usage{
number_vessel_sampled(
  dataframe,
  data_type,
  graph_type = "number",
  reported_year = NULL,
  selected_country = NULL,
  selected_ocean = NULL
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{data_type}{{\link[base]{character}} expected. Tunabio or observe.}

\item{graph_type}{{\link[base]{character}} expected. "number" or "table." Number by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report}

\item{selected_country}{{\link[base]{integer}} expected. Country code to select the list of boat to count. If NULL give all the vessel for the given year.}

\item{selected_ocean}{{\link[base]{integer}} expected. Ocean code to select the list of boat to count. If NULL give all the vessel for the given year.}
}
\value{
The function return a table.
}
\description{
Give the number of vessel sampled for a given year.
}
\details{
The input dataframe frome sql must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  sampling_year}}
 \item{\code{  fish_sampling_date}}
 \item{\code{  landing_date}}
 \item{\code{  vessel_name}}
 \item{\code{  boat_code}}
 \item{\code{  fleet}}
 \item{\code{  vessel_well_number}}
}
add these columns to select the country and ocean (optional):
\itemize{
 \item{\code{  country_id}}
 \item{\code{  ocean_id}}
}
}
