% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_total.R
\name{set_total}
\alias{set_total}
\title{Total set}
\usage{
set_total(
  dataframe,
  graph_type = "table",
  reported_year = NULL,
  selected_country = NULL,
  selected_ocean = NULL,
  selected_harbour = NULL
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the set_total() function.}

\item{graph_type}{{\link[base]{character}} expected. 'number' or 'table'. Number by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report.}

\item{selected_country}{{\link[base]{integer}} expected. Country code to select the list of boat to count. If NULL give all the vessel for the given year.}

\item{selected_ocean}{{\link[base]{integer}} expected. Ocean code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}

\item{selected_harbour}{{\link[base]{integer}} expected. Harbour code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}
}
\value{
The function return a table.
}
\description{
Total sets events, sampled or not, with or without tuna landings.
}
\details{
The input dataframe frome sql must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   program                          | ocean_label | fleet | vessel_label | vessel_type_code | departure  | port_departure | landing_date | country_code | vessel_well_number | arrival   | port_arrival
   -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
   AVDTH Atlantique 1999-2022 (IRD) | Atlantic    | FRA   | VIA AVENIR   | 6                | 2020-12-19 |     ABIDJAN    | 2021-01-15   | 1            | 4T                 | 2021-01-1 | ABIDJAN
   AVDTH Atlantique 1999-2022 (IRD) | Atlantic    | FRA   | VIA AVENIR   | 6                | 2020-12-19 |     ABIDJAN    | 2021-01-15   | 1            | 4T                 | 2021-01-1 | ABIDJAN
   AVDTH Atlantique 1999-2022 (IRD) | Atlantic    | FRA   | VIA AVENIR   | 6                | 2020-12-19 |     ABIDJAN    | 2021-01-15   | 1            | 4T                 | 2021-01-1 | ABIDJAN
}
}
