% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_class_tunabio.R
\name{size_class_tunabio}
\alias{size_class_tunabio}
\title{Representation de l echantillonnage}
\usage{
size_class_tunabio(path_db)
}
\arguments{
\item{path_db}{{\link[base]{character}} expected. Find the path of Tunabio file.}
}
\value{
The function return  ggplot R object.
}
\description{
a completer .
}
