% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_occupancy.R
\name{spatial_occupancy}
\alias{spatial_occupancy}
\title{Spatial occupancy}
\usage{
spatial_occupancy(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_capacity function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in the spatial extent of the fishery over time. Annual number of 1-degree squares explored by each vessel.
}
