% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_serie_catch.R
\name{time_serie_catch}
\alias{time_serie_catch}
\title{Time serie catch}
\usage{
time_serie_catch(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = 1,
  time_serie
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the time_serie_catch() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{time_serie}{{\link[base]{character}} expected. set_fob, set_fsc, catch_fob, catch_fsc}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of sets per searching day and catch per positive set on FOB-associated and free-swimming schools for the French purse seine fishing fleet in the Atlantic Ocean
}
