% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{time_step_control}
\alias{time_step_control}
\title{Control for argument time_step_control}
\usage{
time_step_control(time_step)
}
\description{
Control for argument time_step_control
}
\keyword{internal}
