% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_effort.R
\name{trip_effort}
\alias{trip_effort}
\title{Effort trip}
\usage{
trip_effort(
  dataframe_observe,
  dataframe_t3,
  database = "observe",
  reported_year,
  flag_selected,
  ocean = "Atlantic"
)
}
\arguments{
\item{dataframe_observe}{{\link[base]{data.frame}} expected. Dataframe from the Observe database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the trip_effort() function.}

\item{dataframe_t3}{{\link[base]{data.frame}} expected. Dataframe from the T3 database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the trip_effort() function.}

\item{database}{{\link[base]{character}} expected. Observe or t3. Observe by default.}

\item{reported_year}{{\link[base]{integer}} expected. Year of the report.}

\item{flag_selected}{{\link[base]{character}} expected. Flag of the country; isocode.}

\item{ocean}{{\link[base]{character}} expected. Atlantic or Indian. Atlantic by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Effort trip
}
