% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_number_bis.R
\name{vessel_number_bis}
\alias{vessel_number_bis}
\title{Number of vessel}
\usage{
vessel_number_bis(
  data_connection,
  time_period,
  ocean,
  country,
  vessel_type,
  time_step = "year"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} or name of a internal fishi dataset.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}}. Vessel type codes identification.}

\item{time_step}{{\link[base]{character}} expected. Kind of display you want in the graphic output. You can choose between "month" and "year".}
}
\value{
The function return  ggplot R object.
}
\description{
Number of vessel by country, ocean, period, time step and vessel type.
}
