% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{vessel_type}
\alias{vessel_type}
\alias{vessel_type_control}
\title{Control for argument vessel_type}
\usage{
vessel_type_control(vessel_type)
}
\description{
Control for argument vessel_type
}
\keyword{internal}
