% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessels_number.R
\name{vessels_number}
\alias{vessels_number}
\title{Number of vessels}
\usage{
vessels_number(
  dbconnection,
  countries,
  oceans,
  vessel_types,
  period,
  time_step
)
}
\arguments{
\item{dbconnection}{{\link[base]{list}} or {\link[base]{character}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} or name of a internal fishi dataset.}

\item{countries}{{\link[base]{integer}} expected. Countries codes identifiaction.}

\item{oceans}{{\link[base]{integer}} expected. Oceans codes identifiaction.}

\item{vessel_types}{{\link[base]{integer}}. Vessel types codes identification.}

\item{period}{{\link[base]{integer}} expected. Period identification in year.}

\item{time_step}{{\link[base]{character}} expected. Kind of display you want in the graphic output. You can choose between "month" and "year".}
}
\value{
The function return  ggplot R object.
}
\description{
Number of vessels by country, ocean, period, time step and vessel type.
}
