% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/avdth_nblengthsample_year_month_fleet_ocean_vessel.R
\name{avdth_nblengthsample_year_month_fleet_ocean_vessel}
\alias{avdth_nblengthsample_year_month_fleet_ocean_vessel}
\title{Number of length sampled by year, month, fleet, ocean and vessel (associated to an AVDTH database)}
\usage{
avdth_nblengthsample_year_month_fleet_ocean_vessel(
  avdth_con,
  year,
  fleet,
  ocean,
  fleet_name,
  percentage = F
)
}
\arguments{
\item{avdth_con}{(JDBCConnection object) AVDTH database connection object.}

\item{year}{(integer) Year selected. You can select only one year (related to output design).}

\item{fleet}{(integer) Fleet(s) selected. You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{(integer) Ocean selected. You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fleet_name}{(integer) Fleet(s) name(s).}

\item{percentage}{(logical) If you want to display values in percentages. By default false.}
}
\value{
A ggplot object.
}
\description{
Number of length sampled by year, month, fleet, ocean and vessel (associated to an AVDTH database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean
\dontrun{
tmp <- avdth_nblengthsample_year_month_fleet_ocean(avdth_con = avdth_connection,
                                                   year = 2017,
                                                   fleet = c(1 , 41),
                                                   ocean = 1,
                                                   fleet_name = "french fleet")}
# Same as before but with values in percentages
\dontrun{
tmp <- avdth_nblengthsample_year_month_fleet_ocean(avdth_con = avdth_connection,
                                                   year = 2017,
                                                   fleet = c(1 , 41),
                                                   ocean = 1,
                                                   fleet_name = "french fleet",
                                                   percentage = F)}
}
