% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/balbaya_nbset_fishingmode_gear_year_month_fleet_ocean.R
\name{balbaya_nbset_fishingmode_gear_year_month_fleet_ocean}
\alias{balbaya_nbset_fishingmode_gear_year_month_fleet_ocean}
\title{Number of sets by fishing mode, gear, year, month, fleet and ocean (associated to a balbaya database)}
\usage{
balbaya_nbset_fishingmode_gear_year_month_fleet_ocean(balbaya_con, year,
  fleet, ocean, fishing_mode, gear, fleet_name, monthly = TRUE,
  neg_set = TRUE, pos_set = TRUE, acronym = TRUE)
}
\arguments{
\item{balbaya_con}{Balbaya database connection object.}

\item{year}{Year selected (numerical value). You can select only one year (related to output design).}

\item{fleet}{Fleet(s) selected (numerical value(s)). You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{Ocean selected (numerical value(s)). You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fishing_mode}{Type of fishing mode (numerical value(s)). Check the vignette related to the referentials for more precisely on accepted values.}

\item{gear}{Gear(s) name(s) (numerical value(s)).}

\item{fleet_name}{Fleet(s) name(s) (character value).}

\item{monthly}{If you want to display information monthly (logical value). By default TRUE}

\item{neg_set}{If you want to display information about number of negative set (logical value). By defaut TRUE.}

\item{pos_set}{If you want to display information about number of postive set (logical value). By defaut TRUE.}

\item{acronyme}{If you want to show acronym in the legend or full term. Be default TRUE.}
}
\value{
A R list with data/informations for produce a graphic (stacked area) associated to query data specifications.
}
\description{
Number of sets by fishing mode, year, month, fleet and ocean (associated to a balbaya database).
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean and 2 = Indian Ocean
# For the argument fishing_mode, 1 = Floating object and 2 = Free school
# For the argument gear, 1 = Purse seiner
\dontrun{
tmp1 <- balbaya_nbset_fishingmode_gear_year_month_fleet_ocean(balbaya_con = balabay_connection,
                                                              year = c(2017, 2018),
                                                              fleet = c(1, 41),
                                                              ocean = c(1, 2),
                                                              fishing_mode = c(1, 2),
                                                              gear = 1,
                                                              fleet_name = "French fleet",
                                                              monthly = TRUE,
                                                              neg_set = TRUE,
                                                              pos_set = TRUE,
                                                              acronym = TRUE)}
}
