% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean.R
\name{balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean}
\alias{balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean}
\title{Temporal variables of effort by fishing mode, gear, year, month, fleet and ocean (associated to a balbaya database)}
\usage{

  balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean(balbaya_con,
  year, fleet, ocean, fishing_mode, gear, fleet_name, monthly = TRUE,
  variables = c("time_at_sea", "days_at_sea", "fishing_time",
  "fishing_days", "searching_days"), acronym = TRUE)
}
\arguments{
\item{balbaya_con}{Balbaya database connection object.}

\item{year}{Year selected (numerical value). You can select only one year (related to output design).}

\item{fleet}{Fleet(s) selected (numerical value(s)). You can select several fleets. Check the vignette related to the referentials for more precisely on accepted values.}

\item{ocean}{Ocean selected (numerical value(s)). You can select only one ocean (related to output design). Check the vignette related to the referentials for more precisely on accepted values.}

\item{fishing_mode}{Type of fishing mode (numerical value(s)). Check the vignette related to the referentials for more precisely on accepted values.}

\item{gear}{Gear(s) name(s) (numerical value(s)).}

\item{fleet_name}{Fleet(s) name(s) (character value).}

\item{monthly}{If you want to display information monthly (logical value). By default TRUE}

\item{variables}{List of variables that you want to display in the graphic. By default all variables availabe will display. If you want to know more about variables definitions check the section details below.}

\item{acronyme}{If you want to show acronym in the legend or full term. Be default TRUE.}
}
\value{
A R list with data/informations for produce a graphic (stacked area) associated to query data specifications.
}
\description{
Temporal variables of effort by fishing mode, year, month, fleet and ocean (associated to a balbaya database).
}
\details{
For now, you can display 5 temporal variables:
\itemize{
 \item{time_at_sea: }{time (in hours) of the vessel at sea.}
 \item{days_at_sea: }{time (in days) of the vessel at sea.}
 \item{fishing_time: }{time (in hours) of vessel fishing activity.}
 \item{fishing_days: }{time (in days) of vessel fishing activity. In the Atlantic Ocean a full day of activity is 12 hours by day and 13 hours for the Indian Ocean.}
 \item{searching_days: }{time (in days) of vessel fishing activity (variable fishing_time) less the set duration (time required to make a catch, start when the fishes is incircled).}
}
}
\examples{
# For the argument fleet, 1 = France and 41 = Mayotte
# For the argument ocean, 1 = Atlantic Ocean and 2 = Indian Ocean
# For the argument fishing_mode, 1 = Floating object and 2 = Free school
# For the argument gear, 1 = Purse seiner
tmp1 <- balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean(balbaya_con = balabay_connection,
                                                                        year = c(2017, 2018),
                                                                        fleet = c(1, 41),
                                                                        ocean = c(1, 2),
                                                                        fishing_mode = c(1, 2),
                                                                        gear = 1,
                                                                        fleet_name = "French fleet",
                                                                        monthly = TRUE,
                                                                        variables = c("time_at_sea", "days_at_sea", "fishing_time", "fishing_days", "searching_days"),
                                                                        acronym = TRUE)
}
\references{
\url{https://github.com/OB7-IRD/fishi}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
