% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_weight_tuna.R
\name{bio_weight_tuna}
\alias{bio_weight_tuna}
\title{Weight distribution of major tuna catches}
\usage{
bio_weight_tuna(dataframe, report_year, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the bio_weight_tuna() function.}

\item{report_year}{{\link[base]{integer}} expected. Year of the statistical report.}

\item{graph_type}{{\link[base]{character}} expected. plot or ggplot. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Weight distribution of major tuna catches (in percentage of the total number of fishes).
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  - activity_date}}
 \item{\code{  - c_banc}}
 \item{\code{  - c_esp}}
 \item{\code{  - country_id}}
 \item{\code{  - ocean_id}}
 \item{\code{  - size_class}}
 \item{\code{  - v_mensur}}
}
}
