% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_map.R
\name{catch_map}
\alias{catch_map}
\title{Catch map}
\usage{
catch_map(
  data_connection,
  time_period,
  specie,
  ocean,
  country,
  vessel_type,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the catch_serie function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{specie}{{\link[base]{integer}} expected. Specie codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{path_file}{path to save the final graphic as a png. NULL by default.}
}
\value{
The function return  ggplot R map.
}
\description{
catch_map() graphically represents the distribution of catches on a map by specie, country, ocean, period, time step and vessel type.
}
