% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_per_unit_effort.R
\name{catch_per_unit_effort}
\alias{catch_per_unit_effort}
\title{Annual catch rates (in t per searching day)}
\usage{
catch_per_unit_effort(
  dataframe1,
  dataframe2,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{fishing_type}{{\link[base]{character}} expected. FOB or FSC.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual catch rates (in t per searching day) on FOB- associated and free-swimming tuna schools (FSC).
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{dataframe 1:}
 \item{\code{  - activity_date}}
 \item{\code{  - c_esp}}
 \item{\code{  - country_id}}
 \item{\code{  - c_tban}}
 \item{\code{  - ocean_id}}
 \item{\code{  - v_dur_cal}}
 \item{\code{  - vessel_type_id}}
 \item{\code{  - v_nb_calee_pos}}
 \item{\code{  - v_nb_calees}}
 \item{\code{  - v_poids_capt}}
 \item{\code{  - v_tpec}}
 \item{dataframe 2:}
 \item{\code{  - activity_date}}
 \item{\code{  - country_id}}
 \item{\code{  - c_tban}}
 \item{\code{  - ocean_id}}
 \item{\code{  - v_dur_cal}}
 \item{\code{  - vessel_type_id}}
 \item{\code{  - v_nb_calee_pos}}
 \item{\code{  - v_nb_calees}}
 \item{\code{  - v_tpec}}
}
}
