% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_per_unit_effort.R
\name{catch_per_unit_effort}
\alias{catch_per_unit_effort}
\title{Catch per unit effort}
\usage{
catch_per_unit_effort(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = 1,
  fishing_type
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the catch_per_unit_effort() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{fishing_type}{{\link[base]{character}} expected. FOB or FSC.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual catch rates (in t per searching day) of the French purse seine fishing fleet on FOB- associated and free-swimming tuna schools (FSC) in the Atlantic Ocean.
}
