% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_per_unit_effort.R
\name{catch_per_unit_effort}
\alias{catch_per_unit_effort}
\title{Annual catch rates (in t per searching day)}
\usage{
catch_per_unit_effort(
  dataframe1,
  dataframe2,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the catch_per_searching_day() function.}

\item{fishing_type}{{\link[base]{character}} expected. FOB or FSC.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual catch rates (in t per searching day) on FOB- associated and free-swimming tuna schools (FSC).
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 Dataframe 1:
 \item{\code{  activity_date}}
 \item{\code{  species_code}}
 \item{\code{  school_code}}
 \item{\code{  set_duration}}
 \item{\code{  positive_set}}
 \item{\code{  set_duration}}
 \item{\code{  total_catch_weight}}
 \item{\code{  total_hour_fished}}
}
\itemize{
 Dataframe 2:
 \item{\code{  activity_date}}
 \item{\code{  school_code}}
 \item{\code{  set_duration}}
 \item{\code{  positive_set}}
 \item{\code{  set_duration}}
 \item{\code{  total_hour_fished}}
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
 \item{\code{  vessel_type_code}}
}
}
