% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_serie.R
\name{catch_serie}
\alias{catch_serie}
\title{Catch serie}
\usage{
catch_serie(
  data_connection,
  time_period,
  specie,
  ocean,
  country,
  vessel_type,
  time_step = "year",
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the catch_serie function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year, between 1981 and 2021.}

\item{specie}{{\link[base]{integer}} expected. Specie codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{time_step}{{\link[base]{character}} expected. Kind of display you want in the graphic output. You can choose between "month" and "year".}

\item{path_file}{{\link[base]{character}} expected. NULL by default. Path to save the final graphic as a png.}
}
\value{
The function return ggplot R object.
}
\description{
The purpose of the catch_serie function is to graphically represent the evolution of catches over time by a barplot. It depends on several arguments: the period of time, the species, the oceans, the countries and the types of vessels previously selected by the user. The representation can be done by year or by month depending on the wished result.
}
