% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fleet}
\alias{check_fleet}
\title{Arugment "fleet" verification}
\usage{
check_fleet(fleet)
}
\arguments{
\item{fleet}{(integer) Fleet identification(s)}
}
\description{
Check if the item "fleet" is provided and have a class identified as interger
}
