% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ocean}
\alias{check_ocean}
\title{Arugment "ocean" verification}
\usage{
check_ocean(ocean)
}
\arguments{
\item{ocean}{(integer) Ocean identification}
}
\description{
Check if the item "ocean" is provided and have a unique value identified as class integer
}
