% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_year}
\alias{check_year}
\title{Arugment "year" verification}
\usage{
check_year(year, several_values)
}
\arguments{
\item{year}{(integer) Year(s) selected.}

\item{several_values}{(logical) Allow multiple values or not during checking}
}
\description{
Check if the item "year" is provided and have integer value
}
