% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_trip.R
\name{control_trip}
\alias{control_trip}
\title{Control trip}
\usage{
control_trip(
  dataframe_observe,
  dataframe_t3,
  reported_year,
  flag,
  table = "vessel_set",
  ocean = "Atlantic",
  path_to_csv = NULL
)
}
\arguments{
\item{dataframe_observe}{{\link[base]{data.frame}} expected. Dataframe from the Observe database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the control_trip() function.}

\item{dataframe_t3}{{\link[base]{data.frame}} expected. Dataframe from the T3 database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the control_trip() function.}

\item{reported_year}{{\link[base]{integer}} expected. Year of the report.}

\item{flag}{{\link[base]{character}} expected. Flag of the country; isocode.}

\item{table}{{\link[base]{character}} expected. code_case, controltrip or vessel_set. vessel_set by default.}

\item{ocean}{{\link[base]{character}} expected. Atlantic or Indian. Atlantic by default.}

\item{path_to_csv}{{\link[base]{character}} expected. Path to save the csv file. NULL by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Control trip
}
