% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_trip_map.R
\name{control_trip_map}
\alias{control_trip_map}
\title{Control trip map}
\usage{
control_trip_map(
  dataframe_observe,
  dataframe_t3,
  dataframe_vms,
  trip_i,
  control_dist_vms = FALSE,
  ocean = "Atlantic",
  path_to_shp = NULL
)
}
\arguments{
\item{dataframe_observe}{{\link[base]{data.frame}} expected. Dataframe from the Observe database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{dataframe_t3}{{\link[base]{data.frame}} expected. Dataframe from the T3 database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{dataframe_vms}{{\link[base]{data.frame}} expected. Dataframe from the Vms database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{trip_i}{{\link[base]{character}} expected. Date + # + vessel name.}

\item{control_dist_vms}{TRUE or FALSE. FALSE by default.}

\item{ocean}{{\link[base]{character}} expected. Atlantic or Indian Atlantic by default.}

\item{path_to_shp}{{\link[base]{character}} expected. Put a link, if you want to add a shapefile.}
}
\value{
The function return ggplot R plot.
}
\description{
Map of the control, number of sets, number of fads
}
