% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_availability.R
\name{data_availability}
\alias{data_availability}
\title{Assessing data availability}
\usage{
data_availability(
  dataframe_observe,
  dataframe_t3,
  dataframe_vms,
  graph_type = "ggplot",
  reported_year,
  ocean = "Atlantic"
)
}
\arguments{
\item{dataframe_observe}{{\link[base]{data.frame}} expected. Dataframe from the Observe database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{dataframe_t3}{{\link[base]{data.frame}} expected. Dataframe from the T3 database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{dataframe_vms}{{\link[base]{data.frame}} expected. Dataframe from the Vms database. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the data_availability() function.}

\item{graph_type}{{\link[base]{character}} expected. plot or table. Plot by default.}

\item{reported_year}{{\link[base]{integer}} expected. Year of the report.}

\item{ocean}{{\link[base]{character}} expected. Atlantic or Indian Atlantic by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Control to verify the availability of LB, OBS and OBS data.
}
