% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_production.R
\name{fishery_production}
\alias{fishery_production}
\title{Fishery production}
\usage{
fishery_production(
  data_connection,
  time_period,
  fishing_type = "TOTAL",
  country = as.integer(x = 1),
  ocean = as.integer(x = 1),
  vessel_type = as.integer(x = 1)
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishery_production function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{fishing_type}{{\link[base]{character}} expected. FSC, FOB or TOTAL.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}
}
\value{
The function return ggplot R plot.
}
\description{
Total fishery production. Catch by species of the French purse seine fishing fleet.
}
