% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_production.R
\name{fishery_production}
\alias{fishery_production}
\title{Total fishery production}
\usage{
fishery_production(
  dataframe,
  fishing_type = "ALL",
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishery_production() function.}

\item{fishing_type}{{\link[base]{character}} expected. FSC, FOB or ALL. ALL by default.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly, table or percentage. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Total fishery production (catch by species).
Generates a figure for catches (x 1000 t) of the three main tunas: BET, SKJ and YFT.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  activity_date}}
 \item{\code{  species_code}}
 \item{\code{  flag}}
 \item{\code{  fleet}}
 \item{\code{  gear}}
 \item{\code{  school_code}}
 \item{\code{  ocean_label}}
 \item{\code{  total_catch_weight}}
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
 \item{\code{  vessel_type_code}}
}
}
