% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishi-package.R
\docType{package}
\name{fishi-package}
\alias{fishi}
\alias{fishi-package}
\title{Package FISHeries Indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exploratory and descriptive processes of data associated to fisheries.
}
\section{Effort indicators}{

 \itemize{
 \item{\code{\link{vessels_number}}}
 }
}

\section{Functions for output manipulations}{

\itemize{
\item{\code{\link{code_manipulation}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/fishi}
  \item Report bugs at \url{https://github.com/OB7-IRD/fishi/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} [copyright holder]

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr}
}

}
\keyword{internal}
