% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishi-package.R
\docType{package}
\name{fishi-package}
\alias{fishi}
\alias{fishi-package}
\title{Package FISHeries Indicators}
\description{
Provide statistical indicators for exploratory processes of data associated to fisheries.
}
\section{For AVDTH database}{

\subsection{Catches}{
\itemize{
\item{\code{\link{avdth_nblanding_year_month_fleet_ocean}}}
\item{\code{\link{avdth_catches_sp_fishingmode_year_month_fleet_ocean}}}
\item{\code{\link{avdth_setloc_fishingmode_year_month_fleet_ocean}}}
}}
\subsection{Sampling}{
\itemize{
\item{\code{\link{avdth_wellsample_year_month_fleet_ocean}}}
\item{\code{\link{avdth_nblengthsample_year_month_fleet_ocean}}}
\item{\code{\link{avdth_nblengthsample_year_month_fleet_ocean_vessel}}}
\item{\code{\link{avdth_setlocsample_fishingmode_year_month_fleet_ocean}}}
}}
\subsection{Effort}{
\itemize{
\item{\code{\link{avdth_nbset_fishingmode_year_month_fleet_ocean}}}
}}
}

\section{For balbaya database}{

\subsection{Catches}{
\itemize{
\item{\code{\link{balbaya_catches_sp_fishingmode_year_month_fleet_ocean}}}
}}
\subsection{Effort}{
\itemize{
\item{\code{\link{balbaya_nbset_fishingmode_gear_year_month_fleet_ocean}}}
\item{\code{\link{balbaya_temporal_effort_fishingmode_gear_year_month_fleet_ocean}}}
}}
}

\section{Vignettes link}{

\itemize{
\item{\code{\link{../doc/referentials.html}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/fishi}
  \item Report bugs at \url{https://github.com/OB7-IRD/fishi/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr}

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr}
}

}
\keyword{internal}
