% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishi-package.R
\docType{package}
\name{fishi-package}
\alias{fishi}
\alias{fishi-package}
\title{FISHeries Indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exploratory and descriptive processes of data associated to fisheries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/fishi}
  \item Report bugs at \url{https://github.com/OB7-IRD/fishi/issues}
}

}
\author{
\strong{Maintainer}: Clara Lerebourg \email{clara.lerebourg@ird.fr} (\href{https://orcid.org/0000-0003-0589-8932}{ORCID})

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})
  \item Julien Lebranchu \email{julien.lebranchu@ird.fr} (\href{https://orcid.org/0000-0002-7759-0534}{ORCID})
}

Other contributors:
\itemize{
  \item IRD - Ob7 Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
