% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishi.R
\docType{package}
\name{fishi}
\alias{fishi}
\alias{fishi-package}
\title{Package Fisheries Indicators}
\description{
Provide statistical indicators for exploratory processes of data associated to fisheries.
}
\section{For AVDTH database}{

\itemize{
\item{\code{\link{avdth_nblanding_year_fleet_ocean}}}
\item{\code{\link{avdth_nbset_fishingmode_year_fleet_ocean}}}
\item{\code{\link{avdth_catches_sp_fishingmode_year_fleet_ocean}}}
\item{\code{\link{avdth_wellsample_year_fleet_ocean}}}
\item{\code{\link{avdth_nblengthsample_year_fleet_ocean}}}
\item{\code{\link{avdth_nblengthsample_year_fleet_ocean_vessel}}}
}
}

\section{Vignettes link}{

\itemize{
\item{\code{\link{../doc/referentials.html}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/fishi}
  \item Report bugs at \url{https://github.com/OB7-IRD/fishi/issues}
}

}
