% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_activity.R
\name{fishing_activity}
\alias{fishing_activity}
\title{Annual number of fishing sets}
\usage{
fishing_activity(dataframe, graph_type = "plot", figure = NULL, title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{figure}{{\link[base]{character}} expected. For plotly figure: set (for number of sets graph) or log (for percentage FOB-associated sets graph). NULL by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Fishing operations. Annual number of fishing sets on FOB-associated and free-swimming tuna schools.
}
