% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_activity.R
\name{fishing_activity}
\alias{fishing_activity}
\title{Fishing activity}
\usage{
fishing_activity(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = as.integer(x = 1)
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishing_activity function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}
}
\value{
The function return ggplot R plot.
}
\description{
Fishing operations. Annual number of fishing sets in the French purse seine fishery on FOB- associated and free-swimming tuna schools.
}
