% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_capacity.R
\name{fishing_capacity}
\alias{fishing_capacity}
\title{Fishing capacity}
\usage{
fishing_capacity(
  data_connection,
  time_period,
  ocean,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  graph_type = "plot"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishing_capacity function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{country}{{\link[base]{integer}} expected. Country codes identification. 1 by default.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification. 1 by default.}

\item{graph_type}{{\link[base]{character}} expected. plot or plotly. Plot by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Fishing capacity of the French purse seine fishing fleet in the Atlantic Ocean. Annual changes in the number of purse seiners by tonnage categories (barplots) and total carrying capacity (dashed line with circles).
}
