% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_capacity.R
\name{fishing_capacity}
\alias{fishing_capacity}
\title{Annual changes in the number of vessels}
\usage{
fishing_capacity(dataframe, graph_type = "plot", figure = NULL, title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_capacity() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{figure}{{\link[base]{character}} expected. For plotly figure: vessel (for number of vessels graph) or capacity (for carrying capacity graph). NULL by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Fishing capacity. Annual changes in the number of purse seiners by tonnage categories (barplots) and total carrying capacity (dashed line with circles).
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Referentials.html}{see referentials}]:
\itemize{
 \item{\code{  - activity_date}}
 \item{\code{  - c_quille}}
 \item{\code{  - catch}}
 \item{\code{  - ocean_id}}
 \item{\code{  - vessel_type_id}}
}
}
