% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_effort.R
\name{fishing_effort}
\alias{fishing_effort}
\title{Fishing effort}
\usage{
fishing_effort(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = as.integer(x = 1),
  graph_type = "plot"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fishing_effort function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{graph_type}{{\link[base]{character}} expected. plot or plotly. Plot by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in nominal effort over time. Annual total number of fishing and searching days for the French purse seine fishing fleet in the Atlantic Ocean.
}
