% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_effort.R
\name{fishing_effort}
\alias{fishing_effort}
\title{Annual total number of fishing and searching days}
\usage{
fishing_effort(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_effort() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in nominal effort over time. Annual total number of fishing and searching days.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Referentials.html}{see referentials}]:
\itemize{
 \item{\code{  - ocean_id}}
 \item{\code{  - vessel_type_id}}
 \item{\code{  - country_id}}
 \item{\code{  - fleet}}
 \item{\code{  - flag}}
 \item{\code{  - c_bat}}
 \item{\code{  - l_bat}}
 \item{\code{  - port}}
 \item{\code{  - activity_date}}
 \item{\code{  - landing_date}}
 \item{\code{  - v_tmer}}
 \item{\code{  - v_tpec}}
 \item{\code{  - v_dur_cal}}
}
}
