% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_effort.R
\name{fishing_effort}
\alias{fishing_effort}
\title{Annual total number of fishing and searching days}
\usage{
fishing_effort(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the fishing_effort() function.}

\item{graph_type}{{\link[base]{character}} expected. 'plot', 'plotly' or 'table'. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in nominal effort over time. Annual total number of fishing and searching days.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   activity_date | vessel_code | flag   | fleet  | vessel_label  | landing_date | ocean_code | harbour_label | set_duration | vessel_type_code | total_hour_at_sea | total_hour_fished
   --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
   2010-03-06    | 679         | France | France | MEN GOE       | 2010-03-06   | 1          | ABIDJAN       | 0.0          | 1                |  7.0              |  1.005
   2010-12-04    | 492         | France | France | GUERIDEN      | 2010-12-05   | 1          | TEMA          | 0.0          | 1                | 24.0              | 11.838
   2010-05-19    | 401         | France | France | VIA HARMATTAN | 2010-05-19   | 1          | ABIDJAN       | 0.0          | 1                |  8.0              |  2.048
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
}
}
