% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_time.R
\name{fishing_time}
\alias{fishing_time}
\title{Fleet Operational Hours}
\usage{
fishing_time(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the fishing_time() function.}

\item{graph_type}{{\link[base]{character}} expected. 'plot', 'plotly' or 'table'. Plot by default.}

\item{title}{TRUE or FALSE expected. Title for plotly graph_type. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Represents the time spent by the fleet (in hours)
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   activity_date | fishing_time | gear | country_label | hrsea | vessel_code
   -------------------------------------------------------------------------------
   1991-01-15    |  0           | 1    | FRA           | 24    | 768
   1991-01-19    | 12.0         | 1    | FRA           | 24    | 324
   1991-01-20    |  7.92        | 1    | FRA           | 15    | 324
}

Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
}
}
