% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_effort_distribution.R
\name{map_effort_distribution}
\alias{map_effort_distribution}
\title{Spatial distribution of tuna effort}
\usage{
map_effort_distribution(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the map_effort_distribution() function.}

\item{graph_type}{{\link[base]{character}} expected. plot or plotly. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Spatial distribution of tuna effort.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  - activity_date}}
 \item{\code{  - c_bat}}
 \item{\code{  - country_id}}
 \item{\code{  - cwp11_act}}
 \item{\code{  - n_act}}
 \item{\code{  - ocean_id}}
 \item{\code{  - v_dur_cal}}
 \item{\code{  - vessel_type_id}}
 \item{\code{  - v_tpec}}
}
}
