% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_vessel_sampled.R
\name{number_vessel_sampled}
\alias{number_vessel_sampled}
\title{Vessel sampled}
\usage{
number_vessel_sampled(
  dataframe,
  graph_type = "plot",
  reported_year = NULL,
  selected_country = NULL,
  title = FALSE
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report}

\item{selected_country}{{\link[base]{integer}} expected. Country code to select the list of boat to count}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot or table R plot.
}
\description{
Give the number of vessel sampled for a given year.
}
