% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_summary.R
\name{sample_summary}
\alias{sample_summary}
\title{Sample summary}
\usage{
sample_summary(
  dataframe,
  data_type,
  graph_type = "number",
  reported_year = NULL,
  start_date = NULL,
  end_date = NULL,
  selected_country = NULL,
  selected_ocean = NULL,
  selected_harbour = NULL,
  selected_variable = "trip"
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the sample_summary() function.}

\item{data_type}{{\link[base]{character}} expected. 'tunabio' or 'observe'.}

\item{graph_type}{{\link[base]{character}} expected. 'number' or 'table'. Number by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report.}

\item{start_date}{{\link[base]{character}} expected. if reported_year is not given. Write the start date of the time range of the report.}

\item{end_date}{{\link[base]{character}} expected. if reported_year is not given. Write the end date of the time range of the report}

\item{selected_country}{{\link[base]{integer}} expected. Country code to select the list of boat to count. If NULL give all the vessel for the given year.}

\item{selected_ocean}{{\link[base]{integer}} expected. Ocean code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}

\item{selected_harbour}{{\link[base]{integer}} expected. Harbour code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}

\item{selected_variable}{{\link[base]{character}} expected. Write the variable of the PSU. Can be "trip", "vessel" or "well". "trip" by default.}
}
\value{
The function return a table.
}
\description{
Give the number of trip, vessel or well sampled for a given year. Need to define your goal in the parameter.
}
\details{
The input dataframe frome sql must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  ocean_label}}
 \item{\code{  fleet}}
 \item{\code{  vessel_type}}
 \item{\code{  vessel_label}}
 \item{\code{  landing_year}}
 \item{\code{  departure}}
 \item{\code{  port_departure}}
 \item{\code{  arrival}}
 \item{\code{  port_arrival}}
 \item{\code{  total_landing}}
}
\preformatted{
   ocean_label | fleet | vessel_type_code | vessel_label | landing_year | departure  | port_departure | arrival     | port_arrival | total_landing
   ------------------------------------------------------------------------------------------------------------------------------------------
   Atlantic    | FRA   | 6                | VIA AVENIR   | 2022         | 2022-01-13 | ABIDJAN        | 2022-02-03  | ABIDJAN      |
   Atlantic    | FRA   | 6                | VIA AVENIR   | 2022         | 2022-01-13 | ABIDJAN        | 2022-02-03  | ABIDJAN      |
   Atlantic    | FRA   | 6                | VIA AVENIR   | 2022         | 2022-01-16 | ABIDJAN        | 2022-10-05  | ABIDJAN      |
}
add these columns to select the country and ocean (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
}
}
