% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_per_searching_day.R
\name{set_per_searching_day}
\alias{set_per_searching_day}
\title{Set per searching day}
\usage{
set_per_searching_day(
  data_connection,
  time_period,
  ocean,
  fishing_type,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  graph_type = "plot"
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the set_per_searching_day() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{fishing_type}{{\link[base]{character}} expected. FOB and FSC.}

\item{country}{{\link[base]{integer}} expected. Country codes identification. 1 by default.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification. 1 by default.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of sets per searching day on FOB-associated and free-swimming schools for the French purse seine fishing fleet in the Atlantic Ocean
}
