% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_per_searching_day.R
\name{set_per_searching_day}
\alias{set_per_searching_day}
\title{Set per searching day}
\usage{
set_per_searching_day(
  data_connection,
  time_period,
  country = as.integer(x = 1),
  vessel_type = as.integer(x = 1),
  ocean = 1,
  fishing_type
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the set_per_searching_day() function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{fishing_type}{{\link[base]{character}} expected. FOB and FSC.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of sets per searching day and catch per positive set on FOB-associated and free-swimming schools for the French purse seine fishing fleet in the Atlantic Ocean
}
