% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_per_searching_day.R
\name{set_per_searching_day}
\alias{set_per_searching_day}
\title{Annual number of sets per searching day}
\usage{
set_per_searching_day(
  dataframe,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the set_per_searching_day() function.}

\item{fishing_type}{{\link[base]{character}} expected. FOB and FSC.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of sets per searching day on FOB-associated and free-swimming schools.
}
