% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_per_searching_day.R
\name{set_per_searching_day}
\alias{set_per_searching_day}
\title{Annual number of sets per searching day}
\usage{
set_per_searching_day(
  dataframe,
  fishing_type,
  graph_type = "plot",
  title = FALSE
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. 'Csv' or 'output' of the function {\link[furdeb]{data_extraction}}, which must be done before using the set_per_searching_day() function.}

\item{fishing_type}{{\link[base]{character}} expected. 'FOB' and 'FSC'.}

\item{graph_type}{{\link[base]{character}} expected. 'plot', 'plotly' or 'table.' Plot by default.}

\item{title}{TRUE or FALSE expected. Title for plotly graph_type. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Annual number of sets per searching day on FOB-associated and free-swimming schools.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\preformatted{
   activity_date | school_code | set_duration | positive_set | total_set | total_hour_fished
   -----------------------------------------------------------------------------------------
   2010-03-06    | FOB         | 0            | 0            | 0         |  1.00
   2010-12-04    | FOB         | 0            | 0            | 0         | 11.8
   2010-05-19    | FOB         | 0            | 0            | 0         |  2.05
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_code}}
 \item{\code{  ocean_code}}
 \item{\code{  vessel_type_code}}
}
}
