% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_total.R
\name{set_total}
\alias{set_total}
\title{Total set}
\usage{
set_total(
  dataframe,
  graph_type = "table",
  reported_year = NULL,
  selected_country = NULL,
  selected_ocean = NULL,
  selected_harbour = NULL
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{graph_type}{{\link[base]{character}} expected. "number" or "table." Number by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report.}

\item{selected_country}{{\link[base]{integer}} expected. Country code to select the list of boat to count. If NULL give all the vessel for the given year.}

\item{selected_ocean}{{\link[base]{integer}} expected. Ocean code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}

\item{selected_harbour}{{\link[base]{integer}} expected. Harbour code to select the list of boat to count. If NULL give all the vessel for the given year, works only for 'data_type' == 'observe'}
}
\value{
The function return a table.
}
\description{
Total sets events, sampled or not, with or without tuna landings.
}
\details{
The input dataframe frome sql must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  program}}
 \item{\code{  ocean_name}}
 \item{\code{  fleet}}
 \item{\code{  vessel_type}}
 \item{\code{  vessel_name}}
 \item{\code{  departure}}
 \item{\code{  arrival}}
 \item{\code{  landing_date}}
 \item{\code{  activity_date}}
 \item{\code{  port_departure}}
 \item{\code{  port_arrival}}
}
}
