% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_occupancy.R
\name{spatial_occupancy}
\alias{spatial_occupancy}
\title{Spatial occupancy}
\usage{
spatial_occupancy(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the spatial_occupancy function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Changes in the spatial extent of the fishery over time. Annual number of 1-degree squares explored by each vessel.
}
\details{
The input dataframe must contain all these columns for the function to work [\href{https://ob7-ird.github.io/fishi/articles/Db_and_csv.html}{see referentials}]:
\itemize{
 \item{\code{  activity_date}}
 \item{\code{  cwp11_act}}
 \item{\code{  v_nb_calee_pos}}
 \item{\code{  v_nb_calees}}
 \item{\code{  v_tpec}}
}
Add these columns for an automatic title (optional):
\itemize{
 \item{\code{  country_id}}
 \item{\code{  ocean_id}}
 \item{\code{  vessel_type_id}}
}
}
