% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_occupancy.R
\name{spatial_occupancy}
\alias{spatial_occupancy}
\title{Spatial occupancy}
\usage{
spatial_occupancy(
  data_connection,
  time_period,
  country = as.integer(x = c(1, 41)),
  ocean = as.integer(x = c(1)),
  vessel_type = as.integer(x = c(4, 5, 6))
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the spatial_occupancy function.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification.}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification.}
}
\value{
The function return ggplot R plot.
}
\description{
x
}
