% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_biological_variable.R
\name{species_biological_variable}
\alias{species_biological_variable}
\title{Sampled biological variables}
\usage{
species_biological_variable(dataframe, graph_type = "plot", title = FALSE)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{title}{TRUE or FALSE expected. False by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Give the number of each biological variable sampled for a given year.
}
