% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_biological_variable.R
\name{species_biological_variable}
\alias{species_biological_variable}
\title{Sampled biological variables}
\usage{
species_biological_variable(
  dataframe,
  data_type,
  graph_type = "table",
  reported_year = NULL
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the fishing_activity() function.}

\item{data_type}{{\link[base]{character}} expected. Tunabio or observe.}

\item{graph_type}{{\link[base]{character}} expected. plot or table. table by default.}

\item{reported_year}{{\link[base]{integer}} expected. Write the wanted year of the report}
}
\value{
The function return ggplot or table R plot.
}
\description{
Give the number of each biological variable sampled for a given year.
}
