% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_number.R
\name{vessel_number}
\alias{vessel_number}
\title{Number of vessel}
\usage{
vessel_number(
  dataframe,
  graph_type = "plot",
  figure = NULL,
  title = FALSE,
  time_step = "year"
)
}
\arguments{
\item{dataframe}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[fishi]{data_extraction}}, which must be done before using the vessel_number() function.}

\item{graph_type}{{\link[base]{character}} expected. plot, plotly or table. Plot by default.}

\item{figure}{{\link[base]{character}} expected. For plotly figure: vessel (for number of vessels graph) or capacity (for carrying capacity graph). NULL by default.}

\item{title}{TRUE or FALSE expected. False by default.}

\item{time_step}{{\link[base]{character}} expected. Kind of display you want in the graphic output. You can choose between "month" and "year".}
}
\value{
The function return  ggplot R object.
}
\description{
vessel_number() graphically represents the number of vessel by country, ocean, period, time step and vessel type.
}
